+++ src/core.c
@@ -1395,26 +1395,6 @@ SEXP rnng_strcat(SEXP a, SEXP b) {
 
 }
 
-// weakref ---------------------------------------------------------------------
-
-SEXP rnng_weakref_make(SEXP key, SEXP value) {
-
-  return R_MakeWeakRef(key, value, R_NilValue, FALSE);
-
-}
-
-SEXP rnng_weakref_key(SEXP w) {
-
-  return R_WeakRefKey(w);
-
-}
-
-SEXP rnng_weakref_value(SEXP w) {
-
-  return R_WeakRefValue(w);
-
-}
-
 // next mode selector ----------------------------------------------------------
 
 SEXP rnng_next_mode(SEXP refhook, SEXP mark) {